fix: Fixes in data types for Event

+++ src/main/java/com/business_analytics/model/dto/EventDTO.java
@@ -6,10 +6,10 @@ import java.time.LocalDateTime;
 public class EventDTO {
 
   private String type;
-  private Integer userId;
+  private Long userId;
   private LocalDateTime createdAt;
 
-  public EventDTO(String type, Integer userId, LocalDateTime createdAt) {
+  public EventDTO(String type, Long userId, LocalDateTime createdAt) {
     this.type = type;
     this.userId = userId;
     this.createdAt = createdAt;
@@ -24,11 +24,11 @@ public class EventDTO {
     this.type = type;
   }
 
-  public Integer getUserId() {
+  public Long getUserId() {
     return userId;
   }
 
-  public void setUserId(Integer userId) {
+  public void setUserId(Long userId) {
     this.userId = userId;
   }
